Add more feature support and improve debugging output
Improve stack trace by removing unnecessary output and adding other info
Add get_info() to context which combines file info and the current
node's text to generate more useful context for users.
Print the context stack in a more friendly way on errors and only when
--debug-print is specified
Import gd-impl models from the parent directory by default (template
driven)
Sort imports so they will be generated in a reproducible order.
Fix type of Date/Iso8601Date properties
Change comment_pos to pos, and make it a stack so that we can find
directives in more cases, as well as get the text context to print
diagnostics to the user.