Add for_json(), is_set() and is_initialized() methods
Use the for_json() method to prepare the object for json serialization.
Use the is_set(property_name) method to see if an optional property has
been set.
Use the is_initialized() method  to see if the 'update()' method has
been called. Otherwise the object is not valid.

Also fix some bugs in the template with optional and nullable
properties.