Fix some bugs and add some tests to make sure they stay fixed
Fixes for gd_impl support:
Iso8601Date type was not being used for Date properties.
gd_impl AnyKind for unsupported interface unions was not working.
Neither were being marked as gd_impl so were imported from '.' instead
of '..'

Clarified the property names of the ResolvedType struct and added
verbose comments to avoid future confusion.

Added return types to the template and fixed some issues with the syntax
of for_json()
Prelaunch task for debugging with cargo is unnecessary since it
autobuilds