Fix bugs and add support for Partial built-in and PartialDeep from 'type-fest'
PartialDeep is a very useful utility type that converts all descendants
to partials. We can't do this statically in gdscript but we can flag
descendants as partial_deep and allow them to be partial as well.

Also fix a bug where continue was used instead of break when a non-local
import was encountered, causing it to stop processing the whole file.

Also add support for skipping non-local imports. (imports like
'type-fest' which don't start with a '.' and are usually resolved using
module resolution, which is a bit over-complicated.

Add some support for inheriting from type aliases as well as interfaces.
(needs expansion to cover other test cases that still expect a panic)